adding manual switch via serial
Lots of changes here to make the default valve (generally '0') to be
whatever you want.

* renamed 'updateThruPin() to char2pin(), and it returns a value rather
than change the global variables.

* added pin2char() to do the reverse, which is used for setting the
default values in the state sequence.

* changed the trigger section to check if the 'outPin' (which is in
triggerSequence[]) is the same as the default pin.  This happens when
you initialize the array, or if you run off the end of the sequence.
was setting the pin high and then low very quickly.  Still increments,
it just doesn't change the pin states.